arm64: dts: qcom: msm8996: Add support for the CBF clock
Add the CBF PLL register to the kryocc node and assign a frequency
to the clock.

This makes sure the core cluster interconnect is running at a decent
speed, so that it's no longer a pain to use the device with all cores
enabled.

Since MSM8996 Pro has a different divisor on the CBF PLL, change the
kryocc compatible string in msm8996pro.dtsi to let the driver
know which divisor to use.

Signed-off-by: Konrad Dybcio <konrad.dybcio@somainline.org>
[Yassine: Add MSM8996 Pro apcc compatible]
Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>