slimbus: qcom-ngd-ctrl: Register SSR notifier after all initialization steps
The SSR notifier call can be called as soon as it is registered. This
can cause issues as it accesses at least the ngd_up_work work struct
and qmi_up completion. To make sure nothing uninitialized is used,
move the SSR notifier registration to the end of probe next to PDR
handle allocation, and initialize ngd_up_work before it is registered.

Fixes: a899d324863a3 ("slimbus: qcom-ngd-ctrl: add Sub System Restart support")
Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>