arm64: dts: qcom: msm8996-xiaomi-*: Add LEDs
Add LEDs found on the Xiaomi MSM8996 devices. The devices share
a status RGB LED mounted on the front, as well as a PWM-driven
IR LED for remote control (sometimes known as an IR blaster).
The Mi Note 2 has an additional pair of white LEDs used as backlights
for the touchkeys driven by the PM8994 LPG block.

Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>
Link: https://lore.kernel.org/linux-arm-msm/20220728180120.97968-1-y.oudjana@protonmail.com/T/#u
[msm8996-staging: modified to apply]